+++ CHANGELOG.md
@@ -14,6 +14,10 @@ release date when you use `npm version` (see `README.md`).
 - Add a `fake-update` endpoint to the backend. This can be used to send fake
   delta-chat client generated updates. There is no UI yet.
 
+- A button to remove instances.
+
+- A way to resize some UI sections.
+
 ## [0.15.1][] - 2022-07-04
 
 ### Fixed