+++ CHANGELOG.md
@@ -7,6 +7,8 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+## [0.16.0][] - 2023-02-21
+
 ### Added
 
 - Add a timestamp to messages.
@@ -384,3 +386,7 @@ release date when you use `npm version` (see `README.md`).
 [0.15.0]: https://github.com/webxdc/webxdc-dev/tree/v0.15.0
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.15.1...HEAD
 [0.15.1]: https://github.com/webxdc/webxdc-dev/tree/v0.15.1
+
+
+[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.16.0...HEAD
+[0.16.0]: https://github.com/webxdc/webxdc-dev/tree/v0.16.0